Move bus API and impls to dedicated files
The bus interface can be used by other non-PCI MMIO and PIO devices.
This pave the way for the serial device to use it.
